+++ src/Resource/EnvelopeTemplateRecipient.php
@@ -66,4 +66,6 @@ class EnvelopeTemplateRecipient extends BaseResource
     public bool $approveDocumentsAtOnce;
 
     public string $channelForNotifications;
+
+    public bool $signDocumentsAtOnce;
 }