+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## [Unreleased]
 ### Added
+- Add `CompletedIdentificationReport.name`
 - Add `ReportEndpoint.completedIdentifications` endpoint and `CompletedIdentificationReport` resource
 
 ## [2.3.0] - 2024-03-28