+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## [Unreleased]
 ### Added
+- Add `EnvelopeDocumentsEndpoint.invalidate`
+- Add `EnvelopeDocumentsEndpoint.restore`
 - Add `CompletedIdentificationReport.name`
 - Add `ReportEndpoint.completedIdentifications` endpoint and `CompletedIdentificationReport` resource