+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `IdentificationsEndpoint.updateResult`
 - Add `SignatureScenarioVariant.signDocumentsAtOnce`
 - Add `EnvelopeTagsEndpoint.updateValues` endpoint
+- Add `EnvelopeDocument.assignments` to resource
 
 ## [2.3.0] - 2024-03-28
 ### Added