+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `SignatureScenarioVariant.signDocumentsAtOnce`
 - Add `EnvelopeTagsEndpoint.updateValues` endpoint
 - Add `EnvelopeDocument.assignments` to resource
+- Add `EnvelopeRecipient` signed document columns
 
 ## [2.3.0] - 2024-03-28
 ### Added